Struct cargo_metadata::MetadataCommand

pub struct MetadataCommand { /* private fields */ }

A builder for configurating cargo metadata invocation.

Implementations

impl MetadataCommand

pub fn new() -> MetadataCommand

Creates a default cargo metadata command, which will look for Cargo.toml in the ancestors of the current directory.

pub fn cargo_path(&mut self, path: impl Into<PathBuf>) -> &mut MetadataCommand

Path to cargo executable. If not set, this will use the the $CARGO environment variable, and if that is not set, will simply be cargo.

pub fn manifest_path( &mut self, path: impl Into<PathBuf> ) -> &mut MetadataCommand

Path to Cargo.toml

pub fn current_dir(&mut self, path: impl Into<PathBuf>) -> &mut MetadataCommand

Current directory of the cargo metadata process.

pub fn no_deps(&mut self) -> &mut MetadataCommand

Output information only about the root package and don’t fetch dependencies.

pub fn features(&mut self, features: CargoOpt) -> &mut MetadataCommand

Which features to include.

Call this multiple times to specify advanced feature configurations:

MetadataCommand::new()
    .features(CargoOpt::NoDefaultFeatures)
    .features(CargoOpt::SomeFeatures(vec!["feat1".into(), "feat2".into()]))
    .features(CargoOpt::SomeFeatures(vec!["feat3".into()]))
    // ...

Panics

cargo metadata rejects multiple --no-default-features flags. Similarly, the features() method panics when specifying multiple CargoOpt::NoDefaultFeatures:

MetadataCommand::new()
    .features(CargoOpt::NoDefaultFeatures)
    .features(CargoOpt::NoDefaultFeatures) // <-- panic!
    // ...

The method also panics for multiple CargoOpt::AllFeatures arguments:

MetadataCommand::new()
    .features(CargoOpt::AllFeatures)
    .features(CargoOpt::AllFeatures) // <-- panic!
    // ...

pub fn other_options( &mut self, options: impl Into<Vec<String>> ) -> &mut MetadataCommand

Arbitrary command line flags to pass to cargo. These will be added to the end of the command line invocation.

pub fn cargo_command(&self) -> Command

Builds a command for cargo metadata. This is the first part of the work of exec.

pub fn parse<T: AsRef<str>>(data: T) -> Result<Metadata>

Parses cargo metadata output. data must have been produced by a command built with cargo_command.

pub fn exec(&self) -> Result<Metadata>

Runs configured cargo metadata and returns parsed Metadata.