Struct cargo_metadata::Package

pub struct Package {
    pub name: String,
    pub version: String,
    pub id: String,
    pub dependencies: Vec<Dependency>,
    pub targets: Vec<Target>,
    pub features: HashMap<String, Vec<String>>,
    pub manifest_path: String,
    pub metadata: Value,
    // some fields omitted
}

A crate

Fields
name: String

Name as given in the Cargo.toml

version: String

Version given in the Cargo.toml

id: String

An opaque identifier for a package

dependencies: Vec<Dependency>

List of dependencies of this particular package

targets: Vec<Target>

Targets provided by the crate (lib, bin, example, test, ...)

features: HashMap<String, Vec<String>>

Features provided by the crate, mapped to the features required by that feature.

manifest_path: String

Path containing the Cargo.toml

metadata: Value

Contents of the free form package.metadata section

This contents can be serialized to a struct using serde:

#[macro_use]
extern crate serde_json;
#[macro_use]
extern crate serde_derive;

#[derive(Debug, Deserialize)]
struct SomePackageMetadata {
    some_value: i32,
}

fn main() {
    let value = json!({
        "some_value": 42,
    });

    let package_metadata: SomePackageMetadata = serde_json::from_value(value).unwrap();
    assert_eq!(package_metadata.some_value, 42);
}